requirements.txt
requests>=2.31.0
Pillow>=11.0.0
python-dotenv>=1.0.0
